Animations Added

+++ TodoList/View/ListView.swift
@@ -13,20 +13,28 @@ struct ListView: View {
     
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
-        }
+            else {
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                }
                 .listStyle(PlainListStyle())
+            }
+        }
                 .navigationTitle("ListView  📝")
                 .toolbar {
                     ToolbarItem(placement: .navigationBarLeading) {
